Moved pageLayoutStyles to /src/js/components/Style folder.
