+++ src/js/components/Widgets/IPhoneSpacer.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { hasIPhoneNotch, isIOS, isIOSAppOnMac } from '../../common/utils/cordovaUtils';
 import { isCordova } from '../../common/utils/isCordovaOrWebApp';
 import { renderLog } from '../../common/utils/logging';
-import { IOSNoNotchSpacer, IOSNotchedSpacer } from '../../utils/pageLayoutStyles';
+import { IOSNoNotchSpacer, IOSNotchedSpacer } from '../Style/pageLayoutStyles';
 
 
 // A function component
