+++ src/js/pages/Ballot/Office.jsx
@@ -8,7 +8,7 @@ import IssueActions from '../../actions/IssueActions';
 import OfficeActions from '../../actions/OfficeActions';
 import CandidateList from '../../components/Ballot/CandidateList';
 import LoadingWheel from '../../common/components/Widgets/LoadingWheel';
-import { PageContentContainer } from '../../utils/pageLayoutStyles';
+import { PageContentContainer } from '../../components/Style/pageLayoutStyles';
 import Testimonial from '../../components/Widgets/Testimonial';
 import AppObservableStore from '../../stores/AppObservableStore';
 import BallotStore from '../../stores/BallotStore';
