+++ src/js/pages/More/Attributions.jsx
@@ -4,7 +4,7 @@ import { Link } from 'react-router-dom';
 import attributions from '../../common/constants/attributionText';
 import compileDate from '../../compileDate';
 import { renderLog } from '../../common/utils/logging';
-import { PageContentContainer } from '../../utils/pageLayoutStyles';
+import { PageContentContainer } from '../../components/Style/pageLayoutStyles';
 
 export default class Attributions extends React.Component {
   static getProps () {
