+++ src/js/pages/Process/AppleSignInProcess.jsx
@@ -5,7 +5,7 @@ import { isWebApp } from '../../common/utils/isCordovaOrWebApp';
 import historyPush from '../../common/utils/historyPush';
 import Cookies from '../../common/utils/js-cookie/Cookies';
 import { oAuthLog, renderLog } from '../../common/utils/logging';
-import { PageContentContainer } from '../../utils/pageLayoutStyles';
+import { PageContentContainer } from '../../components/Style/pageLayoutStyles';
 import { stringContains } from '../../utils/textFormat';
 
 export default class AppleSignInProcess extends Component {
