+++ src/js/pages/Settings/ClaimYourPage.jsx
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import Helmet from 'react-helmet';
 import TwitterHandleBox from '../../components/Twitter/TwitterHandleBox';
 import { renderLog } from '../../common/utils/logging';
-import { PageContentContainer } from '../../utils/pageLayoutStyles';
+import { PageContentContainer } from '../../components/Style/pageLayoutStyles';
 
 export default class ClaimYourPage extends Component {
   render () {
