+++ src/js/pages/Settings/HamburgerMenu.jsx
@@ -10,7 +10,7 @@ import VoterStore from '../../stores/VoterStore';
 import { avatarGeneric } from '../../utils/applicationUtils';
 import { isCordova, isWebApp } from '../../common/utils/isCordovaOrWebApp';
 import { renderLog } from '../../common/utils/logging';
-import { PageContentContainer } from '../../utils/pageLayoutStyles';
+import { PageContentContainer } from '../../components/Style/pageLayoutStyles';
 import voterPhoto from '../../common/utils/voterPhoto';
 
 const webAppConfig = require('../../config');
