+++ src/js/pages/Settings/Location.jsx
@@ -11,7 +11,7 @@ import BrowserPushMessage from '../../components/Widgets/BrowserPushMessage';
 import ElectionStore from '../../stores/ElectionStore';
 import VoterStore from '../../stores/VoterStore';
 import { renderLog } from '../../common/utils/logging';
-import { PageContentContainer } from '../../utils/pageLayoutStyles';
+import { PageContentContainer } from '../../components/Style/pageLayoutStyles';
 
 export default class Location extends Component {
   constructor (props) {
